pyproject.toml
[project]
name = "easy-cass-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cassandra-driver>=3.29.2",
    "fastmcp>=2.10.6",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "mypy>=1.17.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "pydantic-settings>=2.8.0",
]
